fix: improve leaderboard refetch flow
This fix removes any lingering information or
inconsistencies in reporting found leaderboard information
to the user.

Before we would call the APIs 2 times whenever we first
entered a map. This no longer happes because of improved
checks.

Now we see what we would expect whenever we back out of
a map and enter a new one, move directly to a new map
in a campaign, or when we enter a new map we have not
played yet.